feat: added some more quick edits
