+++ services/media/workflows/src/Stations/Episodes/EpisodeLicensing/EpisodeLicensing.types.ts
@@ -1,0 +1,5 @@
+import { EpisodesLicensesQuery } from '../../../generated/graphql';
+
+export type EpisodeLicensesData = NonNullable<
+  EpisodesLicensesQuery['episodesLicenses']
+>['nodes'][number];
