+++ services/media/workflows/src/Stations/Ingest/IngestDocumentsExplorer/IngestDocuments.types.ts
@@ -1,0 +1,5 @@
+import { IngestDocumentsQuery } from '../../../generated/graphql';
+
+export type IngestDocumentsData = NonNullable<
+  IngestDocumentsQuery['filtered']
+>['nodes'][number];
