+++ services/media/workflows/src/Stations/Movies/MovieLicensing/MovieLicensing.types.ts
@@ -1,0 +1,5 @@
+import { MoviesLicensesQuery } from '../../../generated/graphql';
+
+export type MovieLicensesData = NonNullable<
+  MoviesLicensesQuery['moviesLicenses']
+>['nodes'][number];
