+++ services/media/workflows/src/Stations/Seasons/SeasonLicensing/SeasonLicensing.types.ts
@@ -1,0 +1,5 @@
+import { SeasonsLicensesQuery } from '../../../generated/graphql';
+
+export type SeasonLicensesData = NonNullable<
+  SeasonsLicensesQuery['seasonsLicenses']
+>['nodes'][number];
