+++ services/media/workflows/src/Stations/TvShows/TvShowLicensing/TvShowLicensing.types.ts
@@ -1,0 +1,5 @@
+import { TvshowsLicensesQuery } from '../../../generated/graphql';
+
+export type TvshowLicensesData = NonNullable<
+  TvshowsLicensesQuery['tvshowsLicenses']
+>['nodes'][number];
