+++ services/media/workflows/src/Stations/TvShows/TvShowLicensingDetails/TvShowLicensingDetailsQuickEdit.tsx
@@ -1,0 +1,23 @@
+import { QuickEditContext, QuickEditContextType } from '@axinom/mosaic-ui';
+import React, { useContext } from 'react';
+import { useParams } from 'react-router-dom';
+import { TvshowLicensesData } from '../TvShowLicensing/TvShowLicensing.types';
+import { TvShowLicensingDetailsForm } from './TvShowLicensingDetailsForm';
+
+interface TvShowLicensingDetailsQuickEditUrlParams {
+  tvshowId: string;
+}
+
+export const TvShowLicensingDetailsQuickEdit: React.FC = () => {
+  const { selectedItem } =
+    useContext<QuickEditContextType<TvshowLicensesData>>(QuickEditContext);
+
+  const { tvshowId } = useParams<TvShowLicensingDetailsQuickEditUrlParams>();
+
+  return (
+    <TvShowLicensingDetailsForm
+      tvshowId={Number(tvshowId)}
+      tvshowsLicenseId={selectedItem.id}
+    />
+  );
+};
